misc updates to scripts for sarix model

+++ fit-sarix-models/combine_forecasts_by_location.R
@@ -1,7 +1,7 @@
 library(tidyverse)
 
 base_path <- "weekly-submission/sarix-forecasts"
-forecast_date <- "2022-04-18"
+forecast_date <- as.character(lubridate::floor_date(Sys.Date(), unit = "week") + 1)
 
 for (target_var in c("cases", "hosps")) {
   target_by_loc_path <- file.path(base_path, paste0(target_var, "-by-loc"))
@@ -9,10 +9,10 @@ for (target_var in c("cases", "hosps")) {
     target_by_loc_path,
     full.names = FALSE,
     recursive = FALSE)
-  models <- c("SARIX_covariates_none_smooth_False_transform_fourthrt_p_7_d_0_P_0_D_0",
-              "SARIX_covariates_none_smooth_False_transform_fourthrt_p_14_d_0_P_0_D_0")
-  models <- c("SARIX_covariates_cases_smooth_False_transform_fourthrt_p_7_d_0_P_0_D_0",
-              "SARIX_covariates_cases_smooth_False_transform_fourthrt_p_14_d_0_P_0_D_0")
+#   models <- c("SARIX_covariates_none_smooth_False_transform_fourthrt_p_7_d_0_P_0_D_0",
+#               "SARIX_covariates_none_smooth_False_transform_fourthrt_p_14_d_0_P_0_D_0")
+#   models <- c("SARIX_covariates_cases_smooth_False_transform_fourthrt_p_7_d_0_P_0_D_0",
+#               "SARIX_covariates_cases_smooth_False_transform_fourthrt_p_14_d_0_P_0_D_0")
 
   for (model in models) {
     files <- list.files(file.path(target_by_loc_path, model))
@@ -38,9 +38,11 @@ for (target_var in c("cases", "hosps")) {
     if (!dir.exists(save_dir)) {
       dir.create(save_dir, recursive = TRUE)
     }
-    readr::write_csv(
-      date_forecasts,
-      file.path(save_dir, paste0(forecast_date, "-", model, ".csv"))
-    )
+    if (!file.exists(file.path(save_dir, paste0(forecast_date, "-", model, ".csv")))) {
+      readr::write_csv(
+        date_forecasts,
+        file.path(save_dir, paste0(forecast_date, "-", model, ".csv"))
+      )
+    }
   }
 }
